Add md5sum checks and timeouts for camApp, to the update script.
If the user didn't safely unmount their flash drive when extracting the
update, it can lead to a corrupted filesystem after applying the update,
so add some md5sums to ensure that the update package is complete
before proceeding.

Some older builds of camApp will rarely hang on SIGKILL, so a timeout is
added to the update script when waiting for camApp to terminate.