Fix packing order of the Raw 12-bit format, add video recovery tool.
The 12-bit packed format had a number of serious bugs in that it would
inconsistently drop the first pixel of a frame, causing an off-by-one error
and colour corruption after debayering, as well as a very confusing bit
packing order. The packing order has been changed to keep the pixels in
the correct order, and to more closely follow the 'y10p'-style of packing.

As an extra bonus, we also add a video memory recovery tool that can be used
to dump the video RAM in the event of a software crash. We hope that no one
ever has a need to use such a tool.