Update Gst/OMX and M3 firmware to fix transparency bug again.
There was a bug in the M3 firmware which could only ever enable the
transparency/colorkeying after starting the frame buffer. Thus if it
ever got enabled then it was impossible to disable. Unfortunately this
happens all the time because the Gst/OMX elements default to enabling
the colorkeying. Thus we fix the bug in the M3 firmware, and also set
colorkey to FALSE by default in Gstreamer.