Fix frame stuttering on new LCDs, add disable-ring-buffer option.
An update to the video daemon was required to fix a frame stuttering
issue seen on the new production LCDs when seeking via the playback
slider. This was due to the touchscreen controller generating update
events significantly faster than the older model.

Also added a option for segmented recording mode to disable the ring
buffer. This allows all segments to be used in this mode rather than
overwriting the last segment before exiting recording mode.