Remove MPEG-TS stream from video system.
The MPEG-TS video stream was found to consume way too much CPU time at
higher bitrates (eg: when the focus peaking is enabled, or there is a
lot of noise). Unfortunately this doesn't seem to be fixable in the
mpegtsmux element, so we will simply remove this feature from Arago
builds of cam-pipeline. The Debian releases will get an RTSP server
instead.